% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-names.R
\docType{data}
\name{common_names}
\alias{common_names}
\title{Nombres comunes en EE.UU}
\format{
U data.frame con 10 flas y 2 columnas
\describe{
\item{name}{El nombre.}
\item{n}{El número total de entradas en la base de datos.}
}
}
\usage{
common_names
}
\description{
Los diez nombres más populares en EE.UU. según el paquete babynames.
}
\keyword{datasets}
