% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{hello_world}
\alias{hello_world}
\title{Saluda}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
hello_world
}
\arguments{
\item{name}{cadena de caracteres con el nombre de la persona a saludar.}

\item{emotion}{cadena de caracteres con la emoción.}
}
\value{
El nombre (\code{name}), invisible.
}
\description{
Te saluda personalmente con emoción.
}
\details{
Imprime en la consola. Las emociones disponibles pueden depender de las
emociones suportadas por tu computadora.
}
\examples{
hello_world("CRAN", emotion = "disbelief")

}
\keyword{datasets}
